fiiiiks

+++ contribution/versions/ezpublish2/trunk/projects/php/ezforum/admin/category.php
@@ -1,6 +1,6 @@
 <?
 /*!
-    $Id: category.php,v 1.12 2000/08/09 10:59:18 jhe-cvs Exp $
+    $Id: category.php,v 1.13 2000/08/09 14:12:44 lw-cvs Exp $
 
     Author: Lars Wilhelmsen <lw@ez.no>
     
@@ -11,13 +11,18 @@
 include( "ezforum/dbsettings.php" );
 include_once( "ezphputils.php" );
 include_once( "template.inc" );
-include_once( "../classes/ezdb.php" );
+//include_once( "../classes/ezdb.php" );
 include_once( "$DOCROOT/classes/ezforumcategory.php" );
-  
+include_once( "../classes/ezusergroup.php" );
+include_once( "../classes/ezsession.php" );
+
+$session = new eZSession();
+
+
 $cat = new eZforumCategory();
 $t = new Template( "$DOCROOT/admin/templates" );
 
-$t->set_file(array( "category" => "category.tpl",
+$t->set_file(Array( "category" => "category.tpl",
                     "category-add" => "category-add.tpl",
                     "category-modify" => "category-modify.tpl",
                     "listelements" => "category-list-elements.tpl"
@@ -25,13 +30,19 @@ $t->set_file(array( "category" => "category.tpl",
 
 $t->set_var( "docroot", $DOCROOT);
 
+if ( $session->get( $AuthenticatedSession ) != 0 )
+{
+    // fail  - reason: user not logged in.
+}
+
 if ( $add )
 {
-    if ( $category_id )
+    if ( !eZUserGroup::verifyCommand( $session->userID, "eZForum_AddCategory" ) )
     {
-        $cat->Id = $category_id;
+        die( "Insufficient user rights" );
+        exit;
     }
-           
+
     $cat->setName( $Name );
     $cat->setDescription( $Description );
     
@@ -45,11 +56,22 @@ if ( $add )
   
 if ($action == "delete")
 {
+    if ( !eZUserGroup::verifyCommand( $session->userID, "eZForum_DeleteCategory" ) )
+    {
+        die( "Insufficient user rights" );
+        exit;
+    }
+        
     $cat->delete( $category_id );
 }
 
 if ( $modifyCategory )
 {
+    if ( !eZUserGroup::verifyCommand( $session->userID(), "eZForum_DeleteCategory" ) )
+    {
+        die( "Insufficient user rights to modify category" );
+        exit;
+    }
     $cat->get( $category_id );
 
     $cat->setName( $Name );
@@ -90,11 +112,10 @@ for ($i = 0; $i < count( $categories ); $i++)
                                                Array("Description", "list-Description" ),
                                                Array("Private", "list-Private" )
                                                )
-                   );
+                 );
 
     $t->set_var( "color", switchColor( $i, "#f0f0f0", "#dcdcdc" ) );
-
-    $t->parse("categories","listelements",true);
+    $t->parse( "categories", "listelements", true );
 }
 
 $t->pparse("output", "category");
